functions/processor-amazon/requirements.txt
functions-framework
google-cloud-firestore
google-cloud-storage
pandas
requests
